requirements.txt
Flask==0.9
Jinja2==2.6
Flask-Script
gears-less
gears-clean-css
gears-coffeescript
Flask-Gears==0.2
cssmin==0.1.4
slimit==0.7.4
wsgiref==0.1.2